Refactor workout queue processors: split into modular files and implement main processor class
- Deleted `worker-workout-processors.ts` and `workout-processors.ts` files.
- Created `workout-processors-main.ts` to serve as the main entry point for workout processing.
- Introduced specialized processors for plans, phases, sessions, and exercises.
- Updated `queue-worker.ts` to utilize the new modular processor structure.
- Added a new script `check-sessions.ts` to query and display session and phase data from the database.
